Crate iceberg_storage_opendal

OpenDAL-based storage implementation for Apache Iceberg.

This crate provides OpenDalStorage and OpenDalStorageFactory, which implement the Storage and StorageFactory traits from the iceberg crate using OpenDAL as the backend.

Structs

AwsCredential

Credential that holds the access_key and secret_key.

CustomAwsCredentialLoader

Custom AWS credential loader. This can be used to load credentials from a custom source, such as the AWS SDK.

Enums

OpenDalStorage

OpenDAL-based storage implementation.

OpenDalStorageFactory

OpenDAL-based storage factory.

Traits

AwsCredentialLoad

Loader trait will try to load credential from different sources.